Move server_util tests into server_util_test.py
